1) Move report outcome to background thread 2) Use object pool for buffer since this can be acessed from multiple threads 3) stringstream reset does not work properly.  replaced with data_buffer 4) use size_t for action id since it indexes an array

+++ reinforcement_learning/rlclientlib/ranking_response_impl.h
@@ -13,8 +13,8 @@ namespace reinforcement_learning {
 
     bool get_choosen_action_id(size_t& action_id) const;
     int set_choosen_action_id(size_t action_id, api_status* status);
-    bool get_action(const size_t idx, int* action_id, float* prob) const;
-    void push_back(const int action_id, const float prob);
+    bool get_action(const size_t idx, size_t* action_id, float* prob) const;
+    void push_back(const size_t action_id, const float prob);
     size_t size() const;
     void set_model_id(const char* model_id);
     const char* get_model_id() const;
@@ -29,7 +29,7 @@ namespace reinforcement_learning {
     std::string _uuid;
     size_t _chosen_action_id;
     std::string _model_id;
-    using coll_type = std::vector<std::pair<int, float>>;
+    using coll_type = std::vector<std::pair<size_t, float>>;
     coll_type _ranking;
     friend class ranking_response;
   };
